requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
asyncio==3.4.3
attrs==24.3.0
billiard==4.2.1
black==24.10.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.4
django-cors-headers==4.6.0
django-ninja==1.3.0
frozenlist==1.5.0
h11==0.14.0
idna==3.10
kombu==5.4.2
loguru==0.7.3
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg==3.2.3
pyasn1==0.6.1
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
rsa==4.9
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
